+++ flow-typed/npm/moment_v2.x.x.js
@@ -1,5 +1,5 @@
-// flow-typed signature: b7d5419a7f10309b2328eeaf6d58f528
-// flow-typed version: b96843401b/moment_v2.x.x/flow_>=v0.28.x
+// flow-typed signature: e23207a33420488837251b6e1897e261
+// flow-typed version: 7822da7258/moment_v2.x.x/flow_>=v0.34.x
 
 type moment$MomentOptions = {
   y?: number|string,
@@ -44,13 +44,15 @@ type moment$MomentCreationData = {
   strict: bool,
 };
 
+type moment$CalendarFormat = string | (moment$Moment) => string;
+
 type moment$CalendarFormats = {
-  sameDay?: string,
-  nextDay?: string,
-  nextWeek?: string,
-  lastDay?: string,
-  lastWeek?: string,
-  sameElse?: string,
+  sameDay?: moment$CalendarFormat,
+  nextDay?: moment$CalendarFormat,
+  nextWeek?: moment$CalendarFormat,
+  lastDay?: moment$CalendarFormat,
+  lastWeek?: moment$CalendarFormat,
+  sameElse?: moment$CalendarFormat,
 };
 
 declare class moment$LocaleData {
@@ -179,8 +181,8 @@ declare class moment$Moment {
   endOf(unit: string): this;
   local(): this;
   utc(): this;
-  utcOffset(offset: number|string): void;
-  utcOffset(): number|string;
+  utcOffset(offset: number|string): this;
+  utcOffset(): number;
   format(format?: string): string;
   fromNow(removeSuffix?: bool): string;
   from(value: moment$Moment|string|number|Date|Array<number>, removePrefix?: bool): string;