+++ src/embeds/creators/amulets.js
@@ -1,8 +1,8 @@
 // @flow
 
-import type { EmbedProps } from './bootstrap';
+import type { EmbedProps } from 'embeds/bootstrap';
 
-import Amulets from './components/Amulets';
+import Amulets from 'embeds/components/Amulets';
 
 export default function (element: HTMLElement, ids: Array<number>) {
   return (props: EmbedProps) => <Amulets {...props} ids={ids} />;