+++ src/features/Character/components/Skill/index.js
@@ -2,6 +2,7 @@
 
 import cx from 'classnames';
 import get from 'lodash/get';
+import T from 'i18n-react';
 
 import TooltipTrigger from 'common/components/TooltipTrigger';
 import Icon from 'common/components/Icon';
@@ -12,13 +13,15 @@ import styles from './styles.less';
 type Props = {
   data?: {},
   className?: string,
+  tooltipTextOverride?: string,
 };
 
-const Skill = ({ data, className }: Props) => {
+const Skill = ({ data, className, tooltipTextOverride }: Props) => {
   const error = get(data, 'error');
+  const tooltipData = tooltipTextOverride || data || T.translate('characters.noSkill');
 
   return (
-    <TooltipTrigger type="skill" data={data || 'No Skill Selected'}>
+    <TooltipTrigger type="skill" data={tooltipData}>
       {(error || !data)
         ? <EmptySkill />
         : (